调试微信支付

+++ lib/pay/wechat/WxPay.php
@@ -102,7 +102,7 @@ class WxPay
 
     private function payInit($amount, $id, $name, $memo)
     {
-        $this->amount       = ceil($amount * 100);//单位为分,所以要乘100
+        $this->amount       = (int)ceil($amount * 100);//单位为分,所以要乘100
         $this->data['id']   = $id;
         $this->data['name'] = $name;
         $this->data['memo'] = $memo;
